container/requirements.txt
spotipy
requests
simplejson
google.cloud
spacy>=3.0.0,<3.1.0
google-api-python-client
google-api-core
google-auth
google-cloud-texttospeech
googleapis-common-protos
graphviz
greenlet
grpcio
gTTS
gTTS-token
google-cloud-language
click==7.1.2
./lib/en_core_web_sm-3.0.0.tar.gz
google-auth
python-dotenv
airflow